#237: add parsing for defining enum values

+++ oc/compiler/symtab/symtab.h
@@ -142,7 +142,7 @@ struct symtab_variable_record_t{
 	//The next hashtable record
 	symtab_variable_record_t* next;
 	//What is the enum member value
-	u_int32_t enum_member_value;
+	u_int64_t enum_member_value;
 	//The current generation of the variable - FOR SSA in CFG
 	u_int16_t current_generation;
 	//The hash of it
